pyproject.toml
[tool.poetry]
name = "gcp-cloudflare-dyndns"
version = "0.1.0"
description = ""
authors = ["Sri Ramanujam"]

[tool.poetry.dependencies]
python = "^3.10"
google-cloud-secret-manager = "^2.8.0"
requests = "^2.27.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
